- 暗記アイテム取得(READ)
